+++ cmd/rpc/server.go
@@ -5,6 +5,18 @@ import (
 	"embed"
 	"encoding/json"
 	"fmt"
+	"io"
+	"io/fs"
+	"net/http"
+	"net/http/pprof"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"strconv"
+	"strings"
+	"sync"
+	"time"
+
 	"github.com/alecthomas/units"
 	"github.com/canopy-network/canopy/controller"
 	"github.com/canopy-network/canopy/fsm"
@@ -21,18 +33,7 @@ import (
 	"github.com/shirou/gopsutil/v3/mem"
 	"github.com/shirou/gopsutil/v3/net"
 	"github.com/shirou/gopsutil/v3/process"
-	"io"
-	"io/fs"
-	"net/http"
-	"net/http/pprof"
-	"os"
-	"os/exec"
 	"path"
-	"path/filepath"
-	"strconv"
-	"strings"
-	"sync"
-	"time"
 )
 
 const (
@@ -69,6 +70,7 @@ const (
 	TxsByRecRouteName              = "txs-by-rec"
 	TxByHashRouteName              = "tx-by-hash"
 	PendingRouteName               = "pending"
+	FailedTxRouteName              = "failed-txs"
 	ProposalsRouteName             = "proposals"
 	PollRouteName                  = "poll"
 	CommitteeRouteName             = "committee"
@@ -167,6 +169,7 @@ var (
 		MinimumEvidenceHeightRouteName: {Method: http.MethodPost, Path: "/v1/query/minimum-evidence-height", HandlerFunc: MinimumEvidenceHeight},
 		LotteryRouteName:               {Method: http.MethodPost, Path: "/v1/query/lottery", HandlerFunc: Lottery},
 		PendingRouteName:               {Method: http.MethodPost, Path: "/v1/query/pending", HandlerFunc: Pending},
+		FailedTxRouteName:              {Method: http.MethodPost, Path: "/v1/query/failed-txs", HandlerFunc: FailedTxs},
 		ProposalsRouteName:             {Method: http.MethodGet, Path: "/v1/gov/proposals", HandlerFunc: Proposals},
 		PollRouteName:                  {Method: http.MethodGet, Path: "/v1/gov/poll", HandlerFunc: Poll},
 		BaseChainInfoRouteName:         {Method: http.MethodPost, Path: "/v1/query/base-chain-info", HandlerFunc: BaseChainInfo},
@@ -335,8 +338,14 @@ func TransactionsByHeight(w http.ResponseWriter, r *http.Request, _ httprouter.P
 }
 
 func Pending(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
-	addrIndexer(w, r, func(_ lib.StoreI, _ crypto.AddressI, p lib.PageParams) (any, lib.ErrorI) {
+	pageIndexer(w, r, func(_ lib.StoreI, _ crypto.AddressI, p lib.PageParams) (any, lib.ErrorI) {
 		return app.GetPendingPage(p)
+	})
+}
+
+func FailedTxs(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
+	addrIndexer(w, r, func(_ lib.StoreI, address crypto.AddressI, p lib.PageParams) (any, lib.ErrorI) {
+		return app.GetFailedTxsPage(address.String(), p)
 	})
 }
 
@@ -1356,6 +1365,24 @@ func addrIndexer(w http.ResponseWriter, r *http.Request, callback func(s lib.Sto
 		write(w, types.ErrAddressEmpty(), http.StatusBadRequest)
 		return
 	}
+	p, err := callback(s, crypto.NewAddressFromBytes(req.Address), req.PageParams)
+	if err != nil {
+		write(w, err, http.StatusBadRequest)
+		return
+	}
+	write(w, p, http.StatusOK)
+}
+
+func pageIndexer(w http.ResponseWriter, r *http.Request, callback func(s lib.StoreI, a crypto.AddressI, p lib.PageParams) (any, lib.ErrorI)) {
+	req := new(paginatedAddressRequest)
+	if ok := unmarshal(w, r, req); !ok {
+		return
+	}
+	s, ok := setupStore(w)
+	if !ok {
+		return
+	}
+	defer s.Discard()
 	p, err := callback(s, crypto.NewAddressFromBytes(req.Address), req.PageParams)
 	if err != nil {
 		write(w, err, http.StatusBadRequest)