+++ fsm/types/error.go
@@ -2,6 +2,7 @@ package types
 
 import (
 	"fmt"
+
 	"github.com/canopy-network/canopy/lib"
 )
 
@@ -293,16 +294,4 @@ func ErrInvalidCheckpoint() lib.ErrorI {
 
 func ErrInvalidStartPollHeight() lib.ErrorI {
 	return lib.NewError(lib.CodeStartPollHeight, lib.StateMachineModule, "start poll height is invalid")
-}
-
-func ErrInvalidNumberOfRewardRecipients() lib.ErrorI {
-	return lib.NewError(lib.CodeInvalidNumberOfRewardRecipients, lib.StateMachineModule, "invalid number of reward recipients")
-}
-
-func ErrInvalidProposerRewardPercent() lib.ErrorI {
-	return lib.NewError(lib.CodeInvalidProposerRewardPercent, lib.StateMachineModule, "invalid proposer reward percent")
-}
-
-func ErrInvalidDelegateReward(address lib.HexBytes, cut uint64) lib.ErrorI {
-	return lib.NewError(lib.CodeInvalidDelegatorReward, lib.StateMachineModule, fmt.Sprintf("invalid delegate reward:\naddress: %s\ncut: %d%%", address, cut))
 }